Add files via upload
Some analysis of the polution data, I extracted all files on a directory called pollution and work there. Change the directory to the folder in your pc where you have extracted the pollution data file. I just made some graphs for one dataset, there's around 400 files. I wonder if we will only do forecasting for one dataset or all of them. If so, I think we can create some maps and use that as our result, a sort of heatmap with the forecastings