Part2: Example data analysis using ARIMA procedure
This script shows using knittr how the process for the analysis of the data set is to be carried. 
To run this script you need to install and activate the following libraries: forecasting, ggplot2, knitr, openair
, timeseries and tseries. Also you need to extract the files of the dataset in a folder and set it as your working directory.


Steps are followed according to flowchart in slides. 
First I worked on the original data points taken every 5 minutes, but then also made some plots and an example forecast by using daily means.

I'm still not clear about the statistics part and how to choose for the arima model. For the example I used the auto arima function to get some parameters and do some temporary forecast for 5 days. 

I hope this would be enough for tomorrow slides.